feat(web): added onboarding page
- feat(web): added completed profile form
- feat(ui): added popover and calendar
- feat(web): added update user profile server action
- feat(db): added birthdate, nickname and bio
- feat(db): added select user by id function

After login users are redirected to onboarding page. Onboarding page checks users login state and it birthdate exists in users table. Also the auth layout checks if user is logged and in which case it also checks if birthdate column has value for the logged in user and redirects to onboarding page accordingly